+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2022-09-05
+### Changed
+- Job proxy checkjob view accepts POST requests
+- Adapt otl interpreter to 1.2.0 otlang version
+### Added
+- Running tests in docker
+
 ## [1.0.5] - 2022-06-20
 ### Fixed
 - Fix `__timestamp__` condition